pyproject.toml
[tool.poetry]
name = "onfido-hometask"
version = "0.1.0"
description = "Googly eyes"
authors = ["Sergei Korneev <korneevsk.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
opencv-python-headless = "^4.10.0.84"
numpy = "^2.1.3"
python-multipart = "^0.0.17"
jinja2 = "^3.1.4"
pydantic = "^2.9.2"
starlette = "^0.41.2"
joblib = "^1.4.2"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
pillow = "^10.0.0"
streamlit = "^1.39.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
